# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the backend for Stitchfolio.",
        "title": "Stitchfolio-backend API docs",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/api/sf/v1",
    "security": [
        {
            "BearerAuth": []
        }
    ],
    "paths": {
...
        "/product": {
            "get": {
                "description": "Get all active products with current stock",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Get all active products",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseModel.DataResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Saves an instance of Product",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Save Product",
                "parameters": [
                    {
                        "description": "product",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestModel.Product"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    }
                }
            }
        },
        "/product/autocomplete": {
            "get": {
                "description": "Autocomplete for products with stock info",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Autocomplete for products",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseModel.ProductAutoComplete"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseModel.DataResponse"
                        }
                    }
                }
            }
        },
        "/product/low-stock": {
            "get": {
                "description": "Get all products with stock below threshold",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Get low stock products",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseModel.DataResponse"
                        }
                    }
                }
            }
        },
        "/product/sku": {
            "get": {
                "description": "Get product details by SKU",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Get product by SKU",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product SKU",
                        "name": "sku",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseModel.DataResponse"
                        }
                    }
                }
            }
        },
        "/product/{id}": {
            "get": {
                "description": "Get an instance of Product with inventory",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Get a specific Product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseModel.DataResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an instance of Product",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Update Product",
                "parameters": [
                    {
                        "description": "product",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestModel.Product"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Product id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes an instance of Product",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Delete Product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "product id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "requestModel.Product": {
            "type": "object",
            "properties": {
                "categoryId": {
                    "type": "integer"
                },
                "costPrice": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "lowStockThreshold": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "sellingPrice": {
                    "type": "number"
                },
                "sku": {
                    "type": "string"
                }
            }
        },
...
        "responseModel.AuditFields": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                }
            }
        },
        "responseModel.Category": {
            "type": "object",
            "properties": {
                "auditFields": {
                    "$ref": "#/definitions/responseModel.AuditFields"
                },
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "productCount": {
                    "description": "Count of products in this category",
                    "type": "integer"
                }
            }
        },
...
        "responseModel.DataResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "pageMetaData": {},
                "type": {
                    "type": "string"
                }
            }
        },
...
        "responseModel.Inventory": {
            "type": "object",
            "properties": {
                "auditFields": {
                    "$ref": "#/definitions/responseModel.AuditFields"
                },
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "isLowStock": {
                    "type": "boolean"
                },
                "lowStockThreshold": {
                    "type": "integer"
                },
                "product": {
                    "description": "Related data",
                    "allOf": [
                        {
                            "$ref": "#/definitions/responseModel.Product"
                        }
                    ]
                },
                "productId": {
                    "type": "integer"
                },
                "productName": {
                    "type": "string"
                },
                "productSku": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "responseModel.Product": {
            "type": "object",
            "properties": {
                "auditFields": {
                    "$ref": "#/definitions/responseModel.AuditFields"
                },
                "category": {
                    "description": "Related data",
                    "allOf": [
                        {
                            "$ref": "#/definitions/responseModel.Category"
                        }
                    ]
                },
                "categoryId": {
                    "type": "integer"
                },
                "categoryName": {
                    "description": "Flattened category name",
                    "type": "string"
                },
                "costPrice": {
                    "type": "number"
                },
                "currentStock": {
                    "description": "From inventory",
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "inventory": {
                    "$ref": "#/definitions/responseModel.Inventory"
                },
                "isActive": {
                    "type": "boolean"
                },
                "isLowStock": {
                    "description": "Stock alert flag",
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "sellingPrice": {
                    "type": "number"
                },
                "sku": {
                    "type": "string"
                }
            }
        },
        "responseModel.ProductAutoComplete": {
            "type": "object",
            "properties": {
                "currentStock": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "isLowStock": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "sku": {
                    "type": "string"
                }
            }
        },
...
        "responseModel.Response": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string"
                },
                "pageMetaData": {},
                "type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Token",
            "in": "header"
        }
    }
}